pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-guitar-accompaniment"
version = "0.1.0"
description = "JamBuddy — AI guitar accompaniment: audio in, backing + lines out"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.23,<2.0",
    "scipy>=1.10",
    "sounddevice>=0.4.6",
    "librosa>=0.10",
    "pyfluidsynth>=1.3",
    "mido>=1.3",
]

[project.optional-dependencies]
dev = ["pytest", "black"]

[project.scripts]
guitar-accompaniment = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "src*"]
